refactor: 모든 사용자의 프로필 드랍다운을 통일
